Complete the ZipappContainer refactoring
* Remove the ZipappContainer interface; it's no longer needed
* Rename ZipappContainerStub to ZipappContainerExtension
